Splitted server code in multiple modules
